feat(chips): Support multiple return values for md-on-append.
The usage of `md-on-append` is not well documented and the
behavior is not consistent.

Fix by updating documentation to set expectations of return
values and updating code to conform to their associated behavior.

Additionally, this adds support for simlultaneously using an
autocomplete selection along with the ability to create new chips.

Previously, the chips directive would not allow for a scenario
which used the autocomplete to provide a list of options, but
also provided a method of inputting new options. The most common
case for this was a tag system which showed existing tags, but
allowed you to create new ones.

Update autocomplete and chips to provide both scenarios and
document how this can be achieved.

Lastly, workaround a few display issues with contact chips demo
(#4450).

_**Note:** This work supercedes PR #3816 which can be closed when
this is merged._

Fixes #4666. Fixes #4193. Fixes #4412. Fixes #4863.